+++ Example/ReusableDemo/CollectionViewCells/MyXIBIndexSquaceCell.swift
@@ -17,8 +17,8 @@ import Reusable
  * and it uses the CollectionView recycling mechanism) => it is `Reusable`
  *
  * That's why it's annotated with the `NibOwnerLoadable` protocol,
- * Which in fact is just a convenience protocol that combines
- * both `NibLoadable` + `Reusable` protocols.
+ * Which in fact is just a typealias that combines
+ * `NibLoadable` & `Reusable` protocols.
  */
 final class MyXIBIndexSquaceCell: UICollectionViewCell, NibReusable {
   @IBOutlet private weak var sectionLabel: UILabel!